chore(endpoints): secure all endpoints with jwt token
- create scripts to create db tables on local and production environment
