+++ routes/index.js
@@ -26,20 +26,36 @@ router.post('/auth/signin',
     AuthController.signin
 );
 
-router.patch('/users/:id/verify', UsersController.verify_user);
-router.get(
-    '/users/:id/account-confirmation', UsersController.confirm_account);
-router.get('/users', UsersController.get_users);
-router.get('/users/:id', UsersController.get_user);
-router.get('/users?status=verified', UsersController.get_users);
+router.patch('/users/:id/verify',
+    AuthenticationMiddleware.verifyToken,
+    UsersController.verify_user
+);
+router.get('/users/:id/account-confirmation',
+    AuthenticationMiddleware.verifyToken,
+    UsersController.confirm_account
+);
+router.get('/users',
+    AuthenticationMiddleware.verifyToken,
+    UsersController.get_users
+);
+router.get('/users/:id',
+    AuthenticationMiddleware.verifyToken,
+    UsersController.get_user
+);
+router.get('/users?status=verified',
+    AuthenticationMiddleware.verifyToken,
+    UsersController.get_users
+);
 router.patch('/users/:id/update',
     UsersValidators.updateProfileValidator,
     UsersController.update_user_profile
 );
-router.get('/users/:id/photo/upload/', 
+router.get('/users/:id/photo/upload/',
+    AuthenticationMiddleware.verifyToken,
     UsersController.get_aws_signed_url
 );
 router.patch('/users/:id/photo/update',
+    AuthenticationMiddleware.verifyToken,
     UsersController.update_photo_url
 );
 router.post('/users/:email/reset_password',
@@ -49,25 +65,50 @@ router.post('/users/:email/reset_password',
 
 router.get('/loans',
     AuthenticationMiddleware.verifyToken,
-    LoansController.get_all_loans);
-router.get('/loans/:id', LoansController.get_loan);
-router.get(
-    '/loans?status=approved&repaid=false', LoansController.get_all_loans);
-router.get(
-    '/loans?status=approved&repaid=true', LoansController.get_all_loans);
+    LoansController.get_all_loans
+);
+router.get('/loans/:id',
+    AuthenticationMiddleware.verifyToken,
+    LoansController.get_loan
+);
+router.get('/loans?status=approved&repaid=false',
+    AuthenticationMiddleware.verifyToken,
+    LoansController.get_all_loans
+);
+router.get('/loans?status=approved&repaid=true',
+    AuthenticationMiddleware.verifyToken,
+    LoansController.get_all_loans
+);
 router.post('/loans',
+    AuthenticationMiddleware.verifyToken,
     LoansValidators.validateAmount,
     LoansValidators.validateTenor,
-    LoansController.create_loan);
-router.patch('/loans/:id/approve', LoansController.approve_or_reject_loan);
-router.patch('/loans/:id/reject', LoansController.approve_or_reject_loan);
-router.get(
-    '/loans/:id/repayments', LoansController.loan_repayment_history
+    LoansController.create_loan
+);
+router.patch('/loans/:id/approve',
+    AuthenticationMiddleware.verifyToken,
+    LoansController.approve_or_reject_loan
+);
+router.patch('/loans/:id/reject',
+    AuthenticationMiddleware.verifyToken,
+    LoansController.approve_or_reject_loan
+);
+router.get('/loans/:id/repayments',
+    AuthenticationMiddleware.verifyToken,
+    LoansController.loan_repayment_history
 );
 router.post('/loans/:id/repayment',
+    AuthenticationMiddleware.verifyToken,
     LoansValidators.validateRepayAmount,
-    LoansController.post_repayment);
-router.get('/repayments', LoansController.get_all_repayments);
-router.get('/repayments/:id', LoansController.get_repayment);
+    LoansController.post_repayment
+);
+router.get('/repayments',
+    AuthenticationMiddleware.verifyToken,
+    LoansController.get_all_repayments
+);
+router.get('/repayments/:id',
+    AuthenticationMiddleware.verifyToken,
+    LoansController.get_repayment
+);
 
 export default router;
