+++ test/users-spec.js
@@ -5,8 +5,7 @@
 import supertest from 'supertest';
 import app from '../app';
 import { test_logger } from '../utils/loggers';
-import createDB from '../utils/createDB';
-import clearDB from '../utils/clearDB';
+import { createDB, clearDB } from '../utils/localDbOps';
 
 const server = supertest.agent(app);
